chore: move ContentView to separate file
